feat: initial version of commitlint plugin for semantic-release
For enforcing the conventional commits style rules, it is beneficial to automate the enforcing as
well. Adding a plugin that checks in the analyzeCommits lifecycle if the commit messages are valid
can prevent merging in and/or publishing commits that do not follow the convention. This plugin
utilizes https://github.com/conventional-changelog/commitlint
(https://github.com/conventional-changelog/commitlint) for checking the messages.